Metrics: add temporality aggregation (#70)
A new Timestamps struct has been added to measurements.zig.
This defines the 2 timestamps detailed in the spec that must be used to signal how to interpret data.

A new component is added in temporality.zig.
This keeps track of data points using their meter or instrument source as an identifier and records their values as well as manages timestamps according to the OTel spec.

Details:

* aggregate with ArrayHashMap
* metrics: add timestamps to data points, read them in OTLP
* refine timestamps data types
* introduce component to apply temporality
* wire temporality in metric reader
* implement .Delta temporality

---------

Signed-off-by: inge4pres <fgualazzi@gmail.com>